refactor(rag): extract detect package
- Move ReuseDetector and SnippetValidator to internal/rag/detect
- Rename files to reuse.go and validator.go
- Resolve dependencies and update references in parent rag package

This is Phase 2.1 of the RAG architectural split.
